[Refactor] #120 - 폴더구조 변경, Model 파일 생성
