+++ PennyPack/Views/MainView/ResultView/ResultModalView.swift
@@ -2,10 +2,12 @@ import SwiftUI
 
 
 struct ResultModalView: View {
+    @StateObject var viewModel: ResultModalViewModel
     @ObservedObject var shoppingViewModel: ShoppingViewModel
     @ObservedObject var listViewModel: ListViewModel
     @State private var isMainViewActive = false
     @Binding var isButton: Bool
+    
     
     var body: some View {
         NavigationStack{
@@ -125,5 +127,5 @@ struct ResultModalView: View {
 }
 
 #Preview {
-    ResultModalView(shoppingViewModel: ShoppingViewModel(),listViewModel: ListViewModel(), isButton: .constant(false))
+    ResultModalView(viewModel: ResultModalViewModel(shoppingViewModel: ShoppingViewModel()), shoppingViewModel: ShoppingViewModel(),listViewModel: ListViewModel(), isButton: .constant(false))
 }
